+++ src/cli/commands/private-cmds/_fetch-cmd.ts
@@ -11,6 +11,7 @@ let compressResponse;
 export default class Fetch implements LegacyCommand {
   name = '_fetch <path> <args>';
   private = true;
+  internal = true;
   description = 'fetch components(s) from a scope';
   alias = '';
   opts = [['n', 'no-dependencies', 'do not include component dependencies']] as CommandOptions;